requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6
pydantic-core==2.14.1

# Database drivers
psycopg2-binary==2.9.9  # PostgreSQL
# mysql-connector-python==8.2.0  # MySQL (optional)
# PyMySQL==1.1.0  # MySQL (optional)

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Security
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
